+++ Project/main.cpp
@@ -13,30 +13,30 @@ int main(int argc, char const *argv[])
     exit(EXIT_FAILURE); 
   }
 
-  int max_cap = atoi(argv[1]);
-  int cap_flr = atoi(argv[2]);
-  int cap_grd = atoi(argv[3]);
-  int cap_off = atoi(argv[4]);
-  int cap_elv = atoi(argv[5]);
-  int num_vst = atoi(argv[6]);
-  int l_circl = atoi(argv[7]);
+  unsigned int max_cap = atoi(argv[1]);
+  unsigned int cap_flr = atoi(argv[2]);
+  unsigned int cap_grd = atoi(argv[3]);
+  unsigned int cap_off = atoi(argv[4]);
+  unsigned int cap_elv = atoi(argv[5]);
+  unsigned int num_vst = atoi(argv[6]);
+  unsigned int l_circl = atoi(argv[7]);
   
   /* Generate visitors required */
   mt19937 randomGen(chrono::steady_clock::now().time_since_epoch().count());      //   Waaaay better than rand()
-  uniform_int_distribution<int> rnd4(1,4),rnd10(1,10);
+  uniform_int_distribution<short> rnd4(1,4),rnd10(1,10);
 
   visitor **ppl = new visitor *[num_vst];
-  for (int i = 0; i < num_vst; ++i)
+  for (unsigned int i = 0; i < num_vst; ++i)
   {
-    int fl  = rnd4(randomGen); 
-    int off = rnd10(randomGen); 
+    short fl  = rnd4(randomGen); 
+    short off = rnd10(randomGen); 
     ppl[i]  = new visitor(fl, off);
   }
 
   /* Create a building, visitors attempt to enter */
   building *service = new building(max_cap, cap_flr, cap_grd, cap_off, cap_elv, l_circl);
 
-  for (int i = 0; i < num_vst; ++i)
+  for (unsigned int i = 0; i < num_vst; ++i)
     service->enter(ppl[i]);           // Get EVERY person in (building->ground_level->waiting_room)
 
   service->get_elevator()->operate();           // Operate 
@@ -44,7 +44,7 @@ int main(int argc, char const *argv[])
   /* Cleanup section */
   delete service;
 
-  for (int i = 0; i < num_vst; ++i)
+  for (unsigned int i = 0; i < num_vst; ++i)
     delete ppl[i];
   delete[] ppl;
